Bugfix wrong type and wrong dereference

+++ obligatory_assignment/include/binary_file.h
@@ -4,7 +4,8 @@
 #include <stdio.h>  // for FILE
 struct Router;
 
-int readBinaryFile(const char* binFile, struct Router** routerArray, size_t* N);
+int readBinaryFile(const char* binFile, struct Router** routerArray,
+                   unsigned int* N);
 int readRouter(FILE* fp, struct Router* router);
 
 #endif  // OBLIGATORY_ASSIGNMENT_INCLUDE_BINARY_FILE_H_
